Add build-kube-cfg
This commit adds a command to build a kubecfg that uses
aws-iam-authenticator.

We want to enable RBAC using aws-iam-authenticator to auth users against
the Kops cluster via IAM roles.

We check if `KOPS_AWS_IAM_AUTHENTICATOR_ENABLED` is set to true, and if
so template our kubecfg. If not set to true, execute the regular
`kops export kubecfg`. Note that we have to fetch the CA data from Kops
S3 state store as we don’t have a kubecfg to pull it from yet.

build-kube-cfg must be run after assuming an AWS role within Geodesic.